I killed the bug which annoyed me about 2 weeks...
The bug happened in test_fib_by_hand when N for fib(N) is small.
The cause was the following.

If the number of children tasks of a parent task reaches to zero,
the last children wakes up the parent.
But I didn't check if the parent is in the state of `taskwait'.
So, if a parent task spawn two children and the first children
finishes its work soon after it is born, it `wakes up' its parent
although the parent task is not in `taskwait' state.

Signed-off-by: Sho Nakatani <lay.sakura@gmail.com>